Enhanced the handling of when a purchase/sale is made Now use the allocateTicket function
FIXED Bug 41 - Now buys the correct number of tickets. When it buys too many, it handles this and is more associated to NEW bug 42
NEW Bug 42 - The agent sometimes buys more than one entertainment ticket in one go
STARTED Bug 42 - The agent handles this by trying to allocate the second ticket. Failing this, it tries to sell it as surplus
FIXED Bug 37,38 - Now handle when a purchase/sale made
FIXED 36 - Seems to have resolved itself and all possible purchase bids are made. Not 100% confirmed